+++ lecture/실전 프론트엔드 테스트 시작하기(by 캡틴판교)/backend/db.json
@@ -181,5 +181,12 @@
       "imageUrl": "https://cdn.pixabay.com/user/2024/03/23/18-39-42-304_250x250.jpg"
     }
   ],
-  "carts": []
+  "carts": [
+    {
+      "id": "0",
+      "name": "Refined Fresh Chicken",
+      "imageUrl": "https://cdn.pixabay.com/user/2016/03/26/22-06-36-459_250x250.jpg",
+      "price": "209.00"
+    }
+  ]
 }
