+++ lecture/실전 프론트엔드 테스트 시작하기(by 캡틴판교)/components/cart/CartHeader.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function CartHeader() {
+  return <h1 data-cy="cart-header">장바구니 페이지</h1>;
+}
