+++ lecture/실전 프론트엔드 테스트 시작하기(by 캡틴판교)/cypress/support/e2e.js
