--- test/실전 프론트엔드 테스트 시작하기(by 캡틴판교)/components/ProductHeader.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function ProductHeader({ title }) {
-  return <h1>{title}</h1>;
-}
