--- test/실전 프론트엔드 테스트 시작하기(by 캡틴판교)/layouts/Layout.jsx
@@ -1,12 +1,0 @@
-import Link from 'next/link';
-
-export default function Layout({ children }) {
-  return (
-    <div>
-      <nav>
-        <Link href="/">상품 목록</Link> | <Link href="/cart">장바구니</Link>
-      </nav>
-      <div>{children}</div>
-    </div>
-  );
-}
